Moved BaseBCFuncEval inheritance to a private class -- I don't actually want users to see this one because it only matters for the interface into multigrid
