+++ src/field_solver/field_solver_multigrid.H
@@ -11,7 +11,6 @@
 #include "field_solver.H"
 #include "mfconductivityopfactory.H"
 #include "dirichlet_func.H"
-#include "CD_ElectrostaticDomainBcFuncEval.H"
 
 #include <AMRMultiGrid.H>
 #include <BiCGStabSolver.H>
